sm changes (#72)
