+++ Cargo.toml
@@ -10,6 +10,7 @@ edition = "2021"
 [dependencies]
 ark-ff = "0.5.0"
 ark-poly = "0.5.0"
+ark-serialize = "0.5.0"
 ark-std ="0.5.0"
 memmap2 = "0.9.5"
